requirements.txt
html2text==2020.1.16
pyperclip==1.8.2
python-dotenv==0.20.0
selenium==4.4.3